applied changes according to the review
