+++ src/app/core/configurable-enum/display-configurable-enum/display-configurable-enum.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 import { DisplayConfigurableEnumComponent } from "./display-configurable-enum.component";
-import { Note } from "../../../../../child-dev-project/notes/model/note";
+import { Note } from "../../../child-dev-project/notes/model/note";
 
 describe("DisplayConfigurableEnumComponent", () => {
   let component: DisplayConfigurableEnumComponent;
