+++ src/app/core/configurable-enum/display-configurable-enum/display-configurable-enum.component.ts
@@ -1,5 +1,5 @@
 import { Component } from "@angular/core";
-import { ViewComponent } from "../view-component";
+import { ViewComponent } from "../../entity-components/entity-utils/view-components/view-component";
 
 /**
  * This component displays a text attribute.
