+++ src/app/core/configurable-enum/edit-configurable-enum/edit-configurable-enum.component.html
