+++ src/app/core/entity-components/entity-list/entity-list.component.ts
@@ -20,10 +20,8 @@ import { Entity, EntityConstructor } from "../../entity/entity";
 import { OperationType } from "../../permissions/entity-permissions.service";
 import { FormFieldConfig } from "../entity-form/entity-form/FormConfig";
 import { EntitySubrecordComponent } from "../entity-subrecord/entity-subrecord/entity-subrecord.component";
-import {
-  FilterComponentSettings,
-  FilterGeneratorService,
-} from "./filter-generator.service";
+import { FilterGeneratorService } from "./filter-generator.service";
+import { FilterComponentSettings } from "./filter-component.settings";
 
 /**
  * This component allows to create a full blown table with pagination, filtering, searching and grouping.
